+++ litmus-portal/frontend/package.json
@@ -22,6 +22,7 @@
     "@material-ui/pickers": "^3.2.10",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.6",
+    "@visx/visx": "^1.7.3",
     "ace-builds": "^1.4.12",
     "brace": "^0.11.1",
     "clsx": "^1.1.1",
@@ -30,6 +31,7 @@
     "d3": "^5.16.0",
     "dagre-d3": "^0.6.4",
     "date-fns": "^2.14.0",
+    "dayjs": "^1.10.4",
     "graphql": "^15.3.0",
     "html2canvas": "^1.0.0-rc.7",
     "i18next": "^19.7.0",
@@ -39,7 +41,7 @@
     "jsonwebtoken": "^8.5.1",
     "jspdf": "^2.1.1",
     "jspdf-autotable": "^3.5.13",
-    "litmus-ui": "^1.1.1",
+    "litmus-ui": "^1.1.2",
     "localforage": "^1.7.3",
     "lodash": "^4.17.20",
     "moment": "^2.27.0",
